Extract nonlinear solver params
