+++ applications/sintering/include/pf-applications/sintering/parameters.h
@@ -18,6 +18,7 @@
 #include <deal.II/base/parameter_handler.h>
 
 #include <pf-applications/lac/solvers_linear_parameters.h>
+#include <pf-applications/lac/solvers_nonlinear_parameters.h>
 
 #include <pf-applications/sintering/preconditioners.h>
 
@@ -303,52 +304,6 @@ namespace Sintering
     bool   output_memory_consumption = false;
   };
 
-  struct NOXData
-  {
-    int         output_information             = 0;
-    std::string direction_method               = "Newton";
-    std::string line_search_method             = "Full Step";
-    std::string line_search_interpolation_type = "Cubic";
-  };
-
-  struct SNESData
-  {
-    std::string solver_name      = "newtonls";
-    std::string line_search_name = "bt";
-  };
-
-  struct NonLinearData
-  {
-    int    nl_max_iter = 10;
-    double nl_abs_tol  = 1.e-20;
-    double nl_rel_tol  = 1.e-5;
-
-    int          l_max_iter       = 1000;
-    double       l_abs_tol        = 1.e-10;
-    double       l_rel_tol        = 1.e-2;
-    std::string  l_solver         = "GMRES";
-    unsigned int l_bisgstab_tries = 30;
-
-    bool         newton_do_update             = true;
-    unsigned int newton_threshold_newton_iter = 100;
-    unsigned int newton_threshold_linear_iter = 20;
-    bool         newton_reuse_preconditioner  = true;
-    bool         newton_use_damping           = true;
-
-    std::string nonlinear_solver_type = "damped";
-
-    bool fdm_jacobian_approximation       = false;
-    bool fdm_precond_system_approximation = false;
-    bool jacobi_free                      = false;
-
-    unsigned int verbosity = 1;
-
-    NOXData                  nox_data;
-    SNESData                 snes_data;
-    LinearSolvers::GMRESData gmres_data;
-  };
-
-
   struct Parameters
   {
     ApproximationData      approximation_data;
@@ -363,7 +318,8 @@ namespace Sintering
     RestartData            restart_data;
     PreconditionersData    preconditioners_data;
     ProfilingData          profiling_data;
-    NonLinearData          nonlinear_data;
+
+    NonLinearSolvers::NonLinearData nonlinear_data;
 
     bool   matrix_based                               = false;
     double grain_cut_off_tolerance                    = 0.0; // 0.00001
