+++ include/pf-applications/lac/solvers_linear_parameters.h
@@ -15,6 +15,8 @@
 
 #pragma once
 
+#include <string>
+
 namespace LinearSolvers
 {
   struct GMRESData
